+++ src/Basket/BasketService.php
@@ -355,4 +355,52 @@ final class BasketService extends AbstractService
 
         return new PaymentInformation($result);
     }
+
+    /**
+     * Add or update comments on the basket or products inside the basket.
+     *
+     * Example:
+     *
+     * $basketService->updateComments($basketId, [
+     *     // Comment on a product inside the basket
+     *     new ProductComment($declinationId, 'Please gift wrap this product.'),
+     *     // Comment on the basket
+     *     new BasketComment('I am superman, please deliver to space'),
+     * ]);
+     *
+     * @param $comments Comment[]
+     * @throws NotFound
+     * @throws SomeParametersAreInvalid
+     */
+    public function updateComments(string $basketId, array $comments): void
+    {
+        $commentsToPost = array_map([self::class, 'serializeComment'], $comments);
+
+        try {
+            $this->client->post(
+                "basket/{$basketId}/comments",
+                [
+                    RequestOptions::FORM_PARAMS => [
+                        'comments' => $commentsToPost,
+                    ],
+                ]
+            );
+        } catch (ClientException $e) {
+            $code = $e->getResponse()->getStatusCode();
+
+            if (404 === $code) {
+                throw new NotFound('Basket not found', $e);
+            }
+            if (400 === $code) {
+                throw new SomeParametersAreInvalid($e->getMessage(), $code, $e);
+            }
+
+            throw $e;
+        }
+    }
+
+    private static function serializeComment(Comment $comment): array
+    {
+        return $comment->toArray();
+    }
 }